Add SASL authentication
Add the possibility to use the SASL authentication to connect to a kafka
cluster.

By default, the SASL authentication is not enabled, the environment
variable `KAFKA_SASL_AUTH_ENABLED` need to be set to `true` and the
`KAFKA_SASL_USER` and `KAFKA_SASL_PASSWORD` need to be set with the
correct user/password.

SASL authentication is required to connect to the Confluent Cloud
clusters.